+++ src/Table.php
@@ -108,8 +108,8 @@ class Table implements Htmlable
         ]);
         if ($validator->fails()) {
             $this->getRequest()->merge([
-                $this->getRowsNumberField() => $this->getRowsNumberValue() ?? config('laravel-table.behavior.rows'),
-                $this->getSearchField() => null,
+                $this->getRowsNumberField() => $this->getRowsNumberValue(),
+                $this->getSearchField() => $this->getRequest()->get($this->getSearchField()),
                 $this->getSortByField() => $this->getSortByValue(),
                 $this->getSortDirField() => $this->getSortDirValue(),
             ]);